requirements.txt
flask
transformers
pillow
werkzeug
torch
